* support for the gig v3 feature to have a number of dimension splits   not equal to a power of two
git-svn-id: https://svn.linuxsampler.org/svn@774 d2c4447f-3961-4fff-bec3-96d84be9abe5

+++ libgig/trunk/src/gig.h
@@ -270,7 +270,7 @@ namespace gig {
         uint8_t      zones;      ///< Number of zones the dimension has.
         split_type_t split_type; ///< Intended for internal usage: will be used to convert a dimension value into the corresponding dimension bit number.
         range_t*     ranges;     ///< Intended for internal usage: Points to the beginning of a range_t array which reflects the value ranges of each dimension zone (only if custom defined ranges are defined, is NULL otherwise).
-        unsigned int zone_size;  ///< Intended for internal usage: reflects the size of each zone (128/zones) for normal split types only, 0 otherwise.
+        float        zone_size;  ///< Intended for internal usage: reflects the size of each zone (128/zones) for normal split types only, 0 otherwise.
     };
 
     /** Defines which frequencies are filtered by the VCF. */
